Removed SFML binaries and added source together with a premake5 script to generate the SFML modules for the project

+++ src/Utility/MessageList.h
@@ -55,7 +55,8 @@ public:
 		m_Packets.clear();
 	}
 
-	int GetSize()const { return m_Packets.size(); }
+	size_t GetSize()const { return m_Packets.size(); }
+
 	static int GetMaxSize() { return MAX_SIZE; }
